requirements.txt
PySide6>=6.6
keyboard>=0.13; platform_system=="Windows"
PyYAML>=6.0
